libhighlight: fix array_index_of key_len type to size_t

+++ subprojects/libhighlight/src/highlight.c
@@ -66,7 +66,7 @@ STATIC_ASSERT(LENGTH(special_capture_names) == HIGHLIGHT_SPECIAL_END);
 
 static int
 array_index_of(
-		const char *const arr[], size_t arr_len, const char *key, int key_len) {
+		const char *const arr[], size_t arr_len, const char *key, size_t key_len) {
 	for (size_t i = 0; i < arr_len; i++) {
 		const char *name = arr[i];
 		if (memcmp(name, key, key_len) == 0 && name[key_len] == '\0') {
